Add Clustering Manager Tab and Changed to Absolute Layout
GroupLayout has a issue adding an JTextArea on this side the window. I
do not know why this is happening, so I am unable to correct what is
wrong. This is why I am changing to absolute layout. It will be easier
to handle and the code size will be smaller. There is no added
difficulty since WindowBuilder handles the code generation for
positioning.